backend/requirements.txt
Django==4.2.21
djangorestframework==3.16.0
psycopg2-binary==2.9.9
django-cors-headers==4.7.0
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.5.0
